+++ frontend/src/css/SuiviMissions.css
@@ -1,0 +1,57 @@
+.app{
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+}
+
+
+.titre {
+    text-align: center;
+    width: 100%;
+    margin-top: 20px;
+}
+
+.missions-list {
+    width: 100%;
+    display: flex; /* Utiliser la mise en page en flexbox */
+    flex-wrap: wrap; /* Permettre aux éléments de passer à la ligne si nécessaire */
+}
+
+.mission {
+    width: 60%;
+    margin: 1% auto;
+    padding: 4%;
+    border: 1px solid #ccc;
+    border-radius: 8px;
+    background-color: #fff;
+    font-family: 'Michroma', sans-serif;
+}
+
+.expanded {
+    /* Styles supplémentaires pour les missions étendues */
+    background-color: #f9f9f9;
+    border: 1px solid #999;
+    padding: 4%;
+}
+
+.valider{
+   background-color: darkgreen;
+    width: 25%;
+    margin-left: 10%;
+    margin-right: 10%;
+}
+
+.abandonner{
+    background-color: darkred;
+    width: 25%;
+    margin-left: 10%;
+    margin-right: 10%;
+}
+
+.valider:hover {
+    background-color: green;
+}
+
+.abandonner:hover {
+    background-color: firebrick;
+}
